+++ mcp_toolkit/mcp_toolkit/CHANGELOG.md
@@ -8,6 +8,23 @@
   - `MCPCallEntry.resource` to create MCPCallEntry for resources.
   - `MCPCallEntry.tool` to create MCPCallEntry for tools.
     This change simplifies the syntax by removing the need to write name of tool twice.
+  - Now `MCPToolDefinition` has inputSchema as required parameter with `ObjectSchema` from `dart_mcp` package for better type safety. For example:
+    ```dart
+      definition: MCPToolDefinition(
+        name: 'calculate_fibonacci',
+        description: 'Calculate the nth Fibonacci number and return the sequence',
+        inputSchema: ObjectSchema(
+          properties: {
+            'n': IntegerSchema(
+              description: 'The position in the Fibonacci sequence (0-100)',
+              minimum: 0,
+              maximum: 100,
+            ),
+          },
+          required: ['n'],
+        ),
+      ),
+    ```
 
 ## 0.1.2